[data] Convert pandas DataFrame to pyarrow Table when creating ArrowBlockAccessor
